refactored TextSink out of HTMLSink
